+++ src/Entity.php
@@ -1,8 +1,6 @@
 <?php
 namespace Penneo\SDK;
 
-use Penneo\SDK\ApiConnector;
-
 abstract class Entity
 {
     protected $id;
@@ -12,23 +10,41 @@ abstract class Entity
     );
     protected static $relativeUrl;
 
+    /**
+     * @param $id
+     *
+     * @return static
+     */
     public static function find($id)
     {
         $class = get_called_class();
         $object = new $class();
         $object->id = $id;
         if (!ApiConnector::readObject($object)) {
-            throw new \Exception('Penneo: Could not find the requested '.$class.' (id = '.$id.')');
+            throw new Exception('Penneo: Could not find the requested '.$class.' (id = '.$id.')');
         }
 
         return $object;
     }
 
+    /**
+     * @return static[]
+     * @throws \Exception
+     */
     public static function findAll()
     {
         return self::findBy(array());
     }
-    
+
+    /**
+     * @param array      $criteria
+     * @param array|null $orderBy
+     * @param null       $limit
+     * @param null       $offset
+     *
+     * @return static[]
+     * @throws \Exception
+     */
     public static function findBy(array $criteria, array $orderBy = null, $limit = null, $offset = null)
     {
         $class = get_called_class();
@@ -36,10 +52,10 @@ abstract class Entity
         // Build query array
         $query = $criteria;
         if ($limit !== null) {
-            $query['limit'] = intval($limit);
+            $query['limit'] = (int) $limit;
         }
         if ($offset !== null) {
-            $query['offset'] = intval($offset);
+            $query['offset'] = (int) $offset;
         }
 
         // Build order by parameters.
@@ -56,7 +72,7 @@ abstract class Entity
 
         $response = ApiConnector::callServer($class::$relativeUrl, null, 'get', array('query' => $query));
         if (!$response) {
-            throw new \Exception('Penneo: Internal problem encountered');
+            throw new Exception('Penneo: Internal problem encountered');
         }
 
         $matches = $response->json();
@@ -70,7 +86,14 @@ abstract class Entity
         
         return $result;
     }
-    
+
+    /**
+     * @param array      $criteria
+     * @param array|null $orderBy
+     *
+     * @return static[]
+     * @throws \Exception
+     */
     public static function findOneBy(array $criteria, array $orderBy = null)
     {
         return self::findBy($criteria, $orderBy, 1);
@@ -96,7 +119,7 @@ abstract class Entity
         }
 
         if (empty($arguments)) {
-            throw \InvalidArgumentException('The method '.$method.$by.' requires parameters');
+            throw new \InvalidArgumentException('The method '.$method.$by.' requires parameters');
         }
 
         $fieldName = lcfirst($by);
@@ -132,27 +155,47 @@ abstract class Entity
         throw new \BadMethodCallException('Unexisting method: '.$method.$by);
     }
 
+    /**
+     * @param Entity $parent
+     * @param        $type
+     * @param        $id
+     *
+     * @return static|false|null
+     */
     public static function findLinkedEntity(Entity $parent, $type, $id)
     {
         $url  = $parent->getRelativeUrl().'/'.$parent->getId().'/'.$type::$relativeUrl.'/'.$id;
 
         $entity = self::getEntity($type, $url, $parent);
         if ($entity === false) {
-            throw new \Exception('Penneo: Internal problem encountered');
+            throw new Exception('Penneo: Internal problem encountered');
         }
 
         return $entity;
     }
 
-    public static function getLinkedEntities(Entity $parent, $type, $url = null)
+    /**
+     * @param Entity $parent
+     * @param string $type      Full class path of the linked entity type
+     * @param null   $url       Force the use of a certain URL instead of the auto-detected one
+     * @param array  $getParams Extra params to be added to the url; must be a associative array of GET params
+     *
+     * @return array|bool
+     * @throws Exception
+     */
+    public static function getLinkedEntities(Entity $parent, $type, $url = null, array $getParams = array())
     {
         if ($url == null) {
-            $url  = $parent->getRelativeUrl().'/'.$parent->getId().'/'.$type::$relativeUrl;
+            $url = $parent->getRelativeUrl().'/'.$parent->getId().'/'.$type::$relativeUrl;
+        }
+
+        if ($getParams) {
+            $url .= '?' . http_build_query($getParams);
         }
 
         $entities = self::getEntities($type, $url, $parent);
         if ($entities === false) {
-            throw new \Exception('Penneo: Internal problem encountered');
+            throw new Exception('Penneo: Internal problem encountered');
         }
 
         return $entities;
@@ -209,7 +252,7 @@ abstract class Entity
 
         $response = ApiConnector::callServer($url, null, 'LINK');
         if (!$response) {
-            throw new \Exception('Penneo: Internal problem encountered');
+            throw new Exception('Penneo: Internal problem encountered');
         }
     
         return true;
@@ -221,7 +264,7 @@ abstract class Entity
 
         $response = ApiConnector::callServer($url, null, 'UNLINK');
         if (!$response) {
-            throw new \Exception('Penneo: Internal problem encountered');
+            throw new Exception('Penneo: Internal problem encountered');
         }
     
         return true;
@@ -233,7 +276,7 @@ abstract class Entity
 
         $response = ApiConnector::callServer($url);
         if (!$response) {
-            throw new \Exception('Penneo: Internal problem encountered fetching assets: '.$assetName);
+            throw new Exception('Penneo: Internal problem encountered fetching assets: '.$assetName);
         }
 
         $assets = $response->json();
@@ -252,7 +295,7 @@ abstract class Entity
         
         $response = ApiConnector::callServer($url, null, 'patch');
         if (!$response) {
-            throw new \Exception('Penneo: Internal problem encountered calling action: '.$actionName);
+            throw new Exception('Penneo: Internal problem encountered calling action: '.$actionName);
         }
             
         return true;
@@ -261,14 +304,14 @@ abstract class Entity
     public static function persist(Entity $object)
     {
         if (!ApiConnector::writeObject($object)) {
-            throw new \Exception('Penneo: Could not persist the '.get_class($object));
+            throw new Exception('Penneo: Could not persist the '.get_class($object));
         }
     }
     
     public static function delete(Entity $object)
     {
         if (!ApiConnector::deleteObject($object)) {
-            throw new \Exception('Penneo: Could not delete the '.get_class($object));
+            throw new Exception('Penneo: Could not delete the '.get_class($object));
         }
         
         $object->id = null;